# Fix compressed length calculations for escaped names