--- a/length_test.go
+++ b/length_test.go
@@ -461,7 +461,7 @@ func TestMsgLengthEscaped(t *testing.T) {
 func TestMsgCompressLengthEscaped(t *testing.T) {
 	msg := new(Msg)
 	msg.Compress = true
-	msg.SetQuestion(`www.example.org.`, TypeA)
+	msg.SetQuestion("www.example.org.", TypeA)
 	msg.Answer = append(msg.Answer, &NS{Hdr: RR_Header{Name: `\000\001\002.example.org.`, Rrtype: TypeNS, Class: ClassINET}, Ns: `ns.\e\x\a\m\p\l\e.org.`})
 	msg.Answer = append(msg.Answer, &NS{Hdr: RR_Header{Name: `www.\e\x\a\m\p\l\e.org.`, Rrtype: TypeNS, Class: ClassINET}, Ns: "ns.example.org."})