*Performance hit incoming.*

`Len` will now take escaping into account and return the correct length regardless of escaping.

These are the benchmarks:
```
name                       old time/op    new time/op    delta
MsgLength-12                  349ns ± 4%     387ns ± 2%   +10.92%  (p=0.000 n=9+9)
MsgLengthNoCompression-12    28.0ns ± 1%    63.3ns ± 2%  +125.91%  (p=0.000 n=10+10)
MsgLengthPack-12             1.10µs ± 8%    1.12µs ± 4%      ~     (p=0.138 n=10+10)
MsgLengthMassive-12          26.1µs ± 2%    27.3µs ± 6%    +4.85%  (p=0.000 n=9+10)
MsgLengthOnlyQuestion-12     10.0ns ± 3%    17.1ns ± 2%   +72.14%  (p=0.000 n=10+9)
MsgLengthEscapedName-12      9.25ns ± 1%   96.08ns ± 1%  +938.58%  (p=0.000 n=9+10)
PackMsg-12                    957ns ± 1%    1015ns ± 1%    +6.10%  (p=0.000 n=10+9)
PackMsgOnlyQuestion-12        173ns ± 2%     182ns ± 2%    +5.08%  (p=0.000 n=9+10)
```

Fixes #841